% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google.R
\name{get_google_translations}
\alias{get_google_translations}
\title{Get translations from Google Translate}
\usage{
get_google_translations(x, lang_to, lang_from = "en",
  google_key = Sys.getenv("GOOGLE_TRANSLATE_API_KEY"),
  parallelization_strategy = c("sequential", "multicore", "cluster"))
}
\arguments{
\item{x}{A character vector containing text to translate.}

\item{lang_to}{A two letter language code describing the language to
translate to. See \code{MICROSOFT_LANGS} for available values.}

\item{lang_from}{A two letter language code describing the language to
translate from, defaulting to English. See \code{MICROSOFT_LANGS} for
available values.}

\item{google_key}{A string containing a subscription key to
Google Translate API. Defaults to \code{Sys.getenv("GOOGLE_TRANSLATE_API_KEY")}.}

\item{parallelization_strategy}{A string naming a parallelization strategy,
passed to \code{\link[future]{plan}}.}
}
\value{
A character vector of translated strings.
}
\description{
Gets translations from Google Translate.
}
\examples{
\donttest{
# Not tested due to need for Google Translate API key
get_google_translations(TRANSLATION_QUOTES, sample(GOOGLE_LANGS, 1))
}
}
\references{
\url{https://cloud.google.com/translate/docs/reference/translate}
}
\seealso{
\code{\link{get_translations}}, \code{\link{get_microsoft_translations}}
}
