% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microsoft.R
\name{get_microsoft_access_token}
\alias{get_microsoft_access_token}
\title{Get Microsoft Translator Access Token}
\usage{

  get_microsoft_access_token(microsoft_key = Sys.getenv("MICROSOFT_TRANSLATOR_API_KEY"))
}
\arguments{
\item{microsoft_key}{A string containing a subscription key to
Microsoft Cognitive Services Translator API.}
}
\value{
A list with two elements.
\describe{
  \item{value}{string containing an access token for Microsoft Cognitive
  Services Translator API, valid for 10 minutes.}
  \item{valid_until}{A \code{POSIXct} time when the token runs out. To give a
  little leeway, this is set to 590 seconds after the token was generated.}
}
}
\description{
Gets an access token to use Microsoft Translator
}
\details{
To get a subscription key, you need to sign up for Microsoft Azure
and subscribe to the Cognitive Services Text API. See
\url{https://www.microsoft.com/en-us/translator/getstarted.aspx}
}
\examples{
\donttest{
# Not tested due to need for Microsoft Cognitive
# Services Translator API key
(access_token <- get_microsoft_access_token())
}
}
\references{
\url{https://docs.microsofttranslator.com/oauth-token.html}
}
