% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microsoft.R
\name{get_microsoft_translations}
\alias{get_microsoft_translations}
\title{Get translations from Microsoft Translator}
\usage{
get_microsoft_translations(x, lang_to, lang_from = "en",
  microsoft_key = Sys.getenv("MICROSOFT_TRANSLATOR_API_KEY"),
  parallelization_strategy = c("sequential", "multicore", "cluster"))
}
\arguments{
\item{x}{A character vector containing text to translate.}

\item{lang_to}{A two letter language code describing the language to
translate to. See \code{MICROSOFT_LANGS} for available values.}

\item{lang_from}{A two letter language code describing the language to
translate from, defaulting to English. See \code{MICROSOFT_LANGS} for
available values.}

\item{microsoft_key}{A string containing a subscription key to
Microsoft Cognitive Services Translator API, passed to
\code{\link{get_microsoft_access_token}}.}

\item{parallelization_strategy}{A string naming a parallelization strategy,
passed to \code{\link[future]{plan}}.}
}
\value{
A character vector of translated strings.
}
\description{
Gets translations from Microsoft Translator.
}
\examples{
\donttest{
# Not tested due to need for Microsoft Cognitive
# Services Translator API key
get_microsoft_translations(TRANSLATION_QUOTES, sample(MICROSOFT_LANGS, 1))
}
}
\references{
\url{https://docs.microsofttranslator.com/text-translate.html#!/default/get_Translate}
}
\seealso{
\code{\link{get_translations}}, \code{\link{get_google_translations}}
}
