% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google.R, R/microsoft.R
\docType{data}
\name{GOOGLE_LANGS}
\alias{GOOGLE_LANGS}
\alias{LANGS}
\alias{MICROSOFT_LANGS}
\title{Languages codes}
\format{An object of class \code{character} of length 81.}
\usage{
GOOGLE_LANGS

MICROSOFT_LANGS
}
\description{
ISO 639-1 two letter langauge codes supported by Google Translate and
Microsoft Translator.
}
\seealso{
\code{\link{get_google_translations}},
\code{\link{get_microsoft_translations}}.
}
\keyword{datasets}
