% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\docType{data}
\name{TRANSLATION_QUOTES}
\alias{TRANSLATION_QUOTES}
\title{Quotes about translations}
\format{A character vector.}
\usage{
TRANSLATION_QUOTES
}
\description{
Some quotes about translating text, for use in the examples.
}
\keyword{datasets}
