% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{get_translations}
\alias{get_translations}
\title{Translate your text, choosing the engine}
\usage{
get_translations(x, lang_to, lang_from = "en", api_key,
  parallelization_strategy = c("sequential", "multicore", "cluster"),
  engine = c("google", "microsoft"))
}
\arguments{
\item{lang_to}{A two letter language code describing the language to
translate to. See \code{MICROSOFT_LANGS} for available values.}

\item{lang_from}{A two letter language code describing the language to
translate from, defaulting to English. See \code{MICROSOFT_LANGS} for
available values.}

\item{parallelization_strategy}{A string naming a parallelization strategy,
passed to \code{\link[future]{plan}}.}

\item{google_key}{A string containing a subscription key to
Google Translate API.}
}
\value{
A character vector of translated strings.
}
\description{
Translate text, choosing whether you use Google Translate or Microsoft
Translator.
#' @param x A character vector containing text to translate.
}
\examples{
\donttest{
# Not tested due to need for Microsoft Cognitive
# Services Translator API key and Google Translate API key
get_translations(
  TRANSLATION_QUOTES,
  "es",
  api_key = Sys.getenv("GOOGLE_TRANSLATE_API_KEY"),
  engine = "google"
)
get_translations(
  TRANSLATION_QUOTES,
  "es",
  api_key = Sys.getenv("MICROSOFT_TRANSLATOR_API_KEY"),
  engine = "microsoft"
)
}
}
\seealso{
\code{\link{get_google_translations}},
\code{\link{get_microsoft_translations}}
}
